update setup.sh
ignores itself

+++ setup.sh
@@ -7,14 +7,16 @@ if [ ${#filenames[@]} -eq 0 ]; then
 else
     for file in "${filenames[@]}"
     do
-        echo "setting up $file"
-        if [ -e ~/.$file ]; then
-            if [ -L ~/.$file ]; then
-                rm ~/.$file
-            else
-                mv ~/.$file ~/.bak
+        if [[ ! $file =~ "setup.sh" ]]; then
+            echo "setting up $file"
+            if [ -e ~/.$file ]; then
+                if [ -L ~/.$file ]; then
+                    rm ~/.$file
+                else
+                    mv ~/.$file ~/.bak
+                fi
             fi
+            ln -s $(dirname $(readlink -f $0))/$file ~/.$file;
         fi
-        ln -s $(dirname $(readlink -f $0))/$file ~/.$file;
     done
 fi
